refactor: refactor import process
